pyproject.toml
[tool.poetry]
name = "javamaint-bench"
version = "0.1.0"
description = "Java Maintenance Agent Benchmark - A→Z Automation"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "jmb", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^2.5.0"
typer = {extras = ["all"], version = "^0.9.0"}
rich = "^13.7.0"
gitpython = "^3.1.40"
pyyaml = "^6.0.1"
pandas = "^2.1.4"
numpy = "^1.26.2"
scikit-learn = "^1.3.2"
matplotlib = "^3.8.2"
jsonschema = "^4.20.0"
tiktoken = "^0.5.2"
httpx = "^0.25.2"
tenacity = "^8.2.3"
jinja2 = "^3.1.2"
openai = "^1.3.7"
anthropic = "^0.7.8"
markdown = "^3.9"
weasyprint = "^66.0"
reportlab = "^4.4.4"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
black = "^23.11.0"
isort = "^5.12.0"
flake8 = "^6.1.0"
mypy = "^1.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
jmb = "jmb.cli:app"
